# openapi.yml
openapi: 3.0.3
info:
  title: TripVote API
  description: みんなの旅投票 (TripVote) – AIが旅行候補を整理し、グループで比較・投票できるサービスのAPI仕様
  version: 1.0.0
  contact:
    name: TripVote Team

servers:
  - url: http://localhost:3000
    description: ローカル開発環境

tags:
  - name: trip-groups
    description: 旅行グループ管理
  - name: candidates
    description: 候補地管理
  - name: questions
    description: 質問管理
  - name: ai
    description: AI機能

paths:
  /api/trip-groups:
    post:
      tags:
        - trip-groups
      summary: グループ作成
      description: 新しい旅行グループを作成します
      operationId: createTripGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: 旅行グループ名
                  example: 夏の沖縄旅行
                departure:
                  type: string
                  maxLength: 200
                  nullable: true
                  description: 出発地点
                  example: 東京
      responses:
        '201':
          description: グループ作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  tripGroup:
                    $ref: '#/components/schemas/TripGroup'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                validation:
                  value:
                    error: グループ名を入力してください
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                serverError:
                  value:
                    error: グループの作成に失敗しました

  /api/trip-groups/{tripGroupId}:
    get:
      tags:
        - trip-groups
      summary: グループ情報取得
      description: 指定されたIDの旅行グループ情報を取得します（候補地は含まれません）
      operationId: getTripGroup
      parameters:
        - name: tripGroupId
          in: path
          required: true
          description: 旅行グループID (UUID)
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: グループ情報取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  tripGroup:
                    $ref: '#/components/schemas/TripGroup'
        '404':
          description: グループが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                notFound:
                  value:
                    error: グループが見つかりません
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                serverError:
                  value:
                    error: エラーが発生しました

  /api/trip-groups/{tripGroupId}/candidates:
    get:
      tags:
        - candidates
      summary: 候補地一覧取得
      description: 指定されたグループの候補地一覧を取得します
      operationId: getCandidates
      parameters:
        - $ref: '#/components/parameters/tripGroupId'
      responses:
        '200':
          description: 候補地一覧取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  candidates:
                    type: array
                    items:
                      $ref: '#/components/schemas/Candidate'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - candidates
      summary: 候補地追加
      description: 旅行グループに新しい候補地を追加します
      operationId: createCandidate
      parameters:
        - $ref: '#/components/parameters/tripGroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                  description: 候補地名
                  example: 美ら海水族館
                source_url:
                  type: string
                  format: uri
                  maxLength: 2000
                  nullable: true
                  description: 情報元URL
                  example: https://example.com/spot/123
      responses:
        '201':
          description: 候補地追加成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  candidate:
                    $ref: '#/components/schemas/Candidate'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                validation:
                  value:
                    error: 候補名を入力してください
        '404':
          description: グループが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                notFound:
                  value:
                    error: グループが見つかりません
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                serverError:
                  value:
                    error: 候補の追加に失敗しました

  /api/trip-groups/{tripGroupId}/questions:
    post:
      tags:
        - questions
      summary: 質問追加
      description: 旅行グループまたは特定の候補地に対して質問を追加します
      operationId: createQuestion
      parameters:
        - $ref: '#/components/parameters/tripGroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  description: 質問内容
                  example: 冬の気候はどうですか？
                candidate_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: 候補地ID (nullの場合は全候補対象)
                  example: 770e8400-e29b-41d4-a716-446655440002
      responses:
        '201':
          description: 質問追加成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  question:
                    $ref: '#/components/schemas/Question'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                validation:
                  value:
                    error: 質問内容を入力してください
        '404':
          description: グループが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                notFound:
                  value:
                    error: グループが見つかりません
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                serverError:
                  value:
                    error: 質問の保存に失敗しました

  /api/ai/question:
    post:
      tags:
        - ai
      summary: AI質問回答 (ストリーミング)
      description: AIが質問に対してストリーミング形式で回答します。回答完了後、DBに保存し候補のai_summary.qaにQ&Aを追記します。
      operationId: answerQuestionWithAI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question_id
                - question
                - trip_group_id
              properties:
                question_id:
                  type: string
                  format: uuid
                  description: 質問ID
                  example: 880e8400-e29b-41d4-a716-446655440003
                question:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  description: 質問内容テキスト
                  example: 冬の気候はどうですか？
                candidate_name:
                  type: string
                  maxLength: 200
                  nullable: true
                  description: 候補地名
                  example: 美ら海水族館
                candidate_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: 候補地ID
                  example: 770e8400-e29b-41d4-a716-446655440002
                trip_group_id:
                  type: string
                  format: uuid
                  description: 旅行グループID
                  example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: AIストリーミング回答（プレーンテキスト）
          content:
            text/plain:
              schema:
                type: string
                description: ReadableStream としてプレーンテキストでトークンがストリーミングされます
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                validation:
                  value:
                    error: パラメータが不正です
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                serverError:
                  value:
                    error: AI回答の生成に失敗しました

  /api/ai/summarize:
    post:
      tags:
        - ai
      summary: AI候補地要約・情報取得
      description: AIが候補地の情報を取得・要約し、候補データを更新して返します
      operationId: summarizeCandidateWithAI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - candidate_id
                - candidate_name
                - trip_group_id
              properties:
                candidate_id:
                  type: string
                  format: uuid
                  description: 候補地ID
                  example: 770e8400-e29b-41d4-a716-446655440002
                candidate_name:
                  type: string
                  minLength: 1
                  maxLength: 200
                  description: 候補地名
                  example: 美ら海水族館
                source_url:
                  type: string
                  format: uri
                  maxLength: 2000
                  nullable: true
                  description: 情報元URL
                  example: https://example.com/spot/123
                trip_group_id:
                  type: string
                  format: uuid
                  description: 旅行グループID
                  example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: AI要約成功（更新された候補データを返却）
          content:
            application/json:
              schema:
                type: object
                properties:
                  candidate:
                    $ref: '#/components/schemas/Candidate'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                validation:
                  value:
                    error: パラメータが不正です
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                serverError:
                  value:
                    error: AI分析に失敗しました

components:
  parameters:
    tripGroupId:
      name: tripGroupId
      in: path
      required: true
      description: 旅行グループID (UUID)
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000

  schemas:
    TripGroup:
      type: object
      properties:
        trip_group_id:
          type: string
          format: uuid
          description: グループID
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: グループ名
          example: 夏の沖縄旅行
        departure:
          type: string
          nullable: true
          description: 出発地点
          example: 東京
        status:
          type: string
          enum: [draft, active, completed]
          description: グループステータス
          example: draft
        created_at:
          type: string
          format: date-time
          description: 作成日時
          example: '2025-01-01T00:00:00.000Z'

    Candidate:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 候補地ID
          example: 660e8400-e29b-41d4-a716-446655440001
        trip_group_id:
          type: string
          format: uuid
          description: 旅行グループID
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: 候補地名
          example: 美ら海水族館
        description:
          type: string
          nullable: true
          description: 説明
          example: 沖縄を代表する大型水族館
        image_url:
          type: string
          format: uri
          nullable: true
          description: 画像URL
          example: https://example.com/image.jpg
        rating:
          type: number
          format: float
          nullable: true
          description: 評価 (0-5)
          example: 4.5
        review_count:
          type: integer
          nullable: true
          description: レビュー数
          example: 1234
        tags:
          type: array
          items:
            $ref: '#/components/schemas/CandidateTag'
          description: タグ
        info:
          type: string
          nullable: true
          description: その他情報
          example: '営業時間: 8:30-18:30'
        ai_summary:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/AISummary'
          description: AI要約
        source_url:
          type: string
          format: uri
          nullable: true
          description: 情報元URL
          example: https://example.com/spot/123
        created_at:
          type: string
          format: date-time
          description: 作成日時
          example: '2025-01-01T00:00:00.000Z'

    CandidateTag:
      type: object
      description: 候補地タグ（アイコン・色情報付き）
      properties:
        icon:
          type: string
          description: アイコン (絵文字)
          example: "\U0001F3D6\uFE0F"
        label:
          type: string
          description: タグ名
          example: 観光
        textColor:
          type: string
          description: テキスト色 (HEX)
          example: '#1E40AF'
        iconColor:
          type: string
          description: アイコン色 (HEX)
          example: '#3B82F6'
        bgColor:
          type: string
          description: 背景色 (HEX)
          example: '#EFF6FF'

    AISummary:
      type: object
      description: AI生成の候補地要約（ヘッドラインとQ&Aリスト）
      properties:
        headline:
          type: string
          description: AI生成のヘッドライン
          example: 沖縄を代表する大型水族館。ジンベエザメが圧巻！
        qa:
          type: array
          items:
            $ref: '#/components/schemas/AISummaryQA'
          description: Q&Aリスト

    AISummaryQA:
      type: object
      description: AI Q&Aペア
      properties:
        q:
          type: string
          description: 質問
          example: おすすめポイントは？
        a:
          type: string
          description: 回答
          example: ジンベエザメが泳ぐ世界最大級の水槽が圧巻です

    Question:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 質問ID
          example: 880e8400-e29b-41d4-a716-446655440003
        trip_group_id:
          type: string
          format: uuid
          description: 旅行グループID
          example: 550e8400-e29b-41d4-a716-446655440000
        candidate_id:
          type: string
          format: uuid
          nullable: true
          description: 候補地ID
          example: 770e8400-e29b-41d4-a716-446655440002
        content:
          type: string
          description: 質問内容
          example: 冬の気候はどうですか？
        ai_answer:
          type: string
          nullable: true
          description: AI回答
        created_at:
          type: string
          format: date-time
          description: 作成日時
          example: '2025-01-01T00:00:00.000Z'

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: エラーメッセージ
          example: グループ名を入力してください